Move user entities and user utility functions to separate files
