+++ src/entities/index.ts
@@ -1,0 +1,1 @@
+export { nullUser, exampleUsers } from './user';
