+++ src/utilities/index.ts
@@ -2,3 +2,4 @@ export { checkSignUpStep } from './check-sign-up-step';
 export { fitFetch, fitFetcher } from './fetch';
 export { getStorageUrl, uploadProfileImage } from './storage';
 export { getTokens, setTokens } from './session';
+export { isUserStudent, isUserWorker } from './user';
